anvi_profile rule now has the profile.db as output and not a directory
By adding `--overwrite-output-destinations` to the arguments of `
anvi-profile`, we can now set the profile database as the output of the
rule. The reason we need to add the `--overwrite-output-destinations`
flag is because snakemake creates the output folder automatically and
anvio will not like that. It's better to have the profile database as
output because snakemake is meant to track files and not folders.