apply `ancient` to `path-to-raw-fastq-files.txt`
Basically rejecting the premise of 633dccf, since if I will allow a new
`.ini` file created during a snakemake run then the `ancient` that is
applied to the `.ini` file (in the `qc` rule) will not do anything and
`qc` would run on all samples. Hence, we need to apply ancient on the
input to `gen_configs`.